refactor huge mono-file into smaller modules after mouse wheel broke from too much scrolling
